+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - cmsdev: Limit redundant logging and output related to KUBECONFIG environment variable
 - cmsdev: Update VCS test to reflect change to logical DB backups in CSM 1.5
+- cmsdev: Modify BOS test to handle change to Bitnami for etcd; make pod checks more flexible
 
 ## [1.11.7] - 2023-04-04